requirements.txt
torch~=2.8.0
torchvision~=0.23.0
pillow~=11.3.0
pandas~=2.3.1
matplotlib~=3.10.3
tqdm~=4.67.1
torchmetrics~=1.8.0
numpy~=2.2.6
scipy~=1.16.1
sympy~=1.14.0